+++ src/module/MPC/AdaptiveMPC/Optimizer.cpp
@@ -10,7 +10,7 @@ namespace MPC {
     namespace AdaptiveMPC {
         Optimizer optimizer1 = Optimizer(40, 0, 0);
 
-        Optimizer::Optimizer(int ch_max, double state_weight, double input_weight)
+        Optimizer::Optimizer(int ch_max, float state_weight, float input_weight)
             : ch_max(ch_max), state_weight(state_weight), input_weight(input_weight) {}
     }  // namespace AdaptiveMPC
 }  // namespace MPC
